+++ src/sll/platform/posix/posix_console.c
@@ -2,6 +2,11 @@
 
 
 
+__SLL_EXTERNAL void sll_platform_create_console(void){
+}
+
+
+
 __SLL_EXTERNAL void sll_platform_enable_console_color(void){
 }